README.md
Updating Kaua'i Nui Kuapapa Website

The purpose of this project will be to update the current site informing the public of Kaua'i's historical Moku and Ahupuaʻa land management system, to create a better learning experience for users. I would love to make this project happen and be apart of making Hawaiian history more accessible to all. I'm also interested in this project because I believe it will require a well understanding of all that we have been taught so far. 

This project will include:
An interactive map of Kaua'i and Ni'ihau as the Home Page:
Map will be divided by Moku, hover over and select a Moku to reveal the Ahupua'as within the Moku 
By creating this map you are able to select a Moku from a map of Kaua'i instead of currently selecting from a list of Moku names in which a user may not be familiar with yet.
Features that will be added to the site :
Surf + Weather Report based on user location 
Water Safety info/video
Share Your Mana'o contact/email section

My next step is to create some mock ups and contact the group with the proposal.
Potential road blocks could happen when adding new features, like the weather based on user location. Also, with the interactive map I think could be a challenge to determine where the click event coordinates are for each Moku.

I plan to spend 10-15 hours a week on this project and to be completed after 5-7 weeks.